+++ WebServicesThermostatBackend/Service/ThermostatService.cs
@@ -6,10 +6,10 @@ namespace ThermostatBackend.Service
     public class ThermostatService
     {
 
-        public static Tuple<bool,float> updateTemperature(float temperature)
+        public static Tuple<bool, double> updateTemperature(double temperature)
         {
             var result = ThermostatModel.setTemperature(temperature);
-            return (result.Item1) ? result: new Tuple<bool, float>(false, float.NaN);
+            return (result.Item1) ? result: new Tuple<bool, double>(false, double.NaN);
         }
 
         public static Tuple<bool, int?> setID(int newId)
@@ -24,10 +24,10 @@ namespace ThermostatBackend.Service
             return (result.Item1) ? result : new Tuple<bool, int?>(false, null);
         }
 
-        public static Tuple<bool, float> getTemperature()
+        public static Tuple<bool, double> getTemperature()
         {
             var result = ThermostatModel.getTemperature();
-            return (result.Item1) ? result : new Tuple<bool, float>(false, float.NaN);
+            return (result.Item1) ? result : new Tuple<bool, double>(false, double.NaN);
         }
     }
 }
